pyproject.toml
[tool.poetry]
name = "Hera"
version = "2.2.0"
description = "An unoffical telegram bot for SIMConnect timetables."

license = "AGPL-3.0-only"

authors = [
    "Tan Jing Kai <contact@jingk.ai>"
]

readme = 'README.md'  # Markdown files are supported

repository = "https://github.com/xlanor/SIM-UoW-Timetable-bot"
homepage = "https://github.com/xlanor/SIM-UoW-Timetable-bot"

keywords = ['telegram bot', 'SIMConnect', 'web scraping']

[tool.poetry.dependencies]
python = "^3.6" 
toml = "^0.9"
# Dependencies with extras
selenium = { version = "3.10.0" }
pip = { version = "~9.0.1" }
redis = { version = "~2.10.6" }
jsonpickle = { version = "~0.9.6" }
pymongo = { version = "~3.7.1" }
python_telegram_bot = { version = "~11.1.0" }
celery = { version = "4.1.1" }
beautifulsoup4 = { version = "4.6.3" }
typing = { version = "3.6.6" }

[tool.poetry.dev-dependencies]
pytest = "^3.8.0"
pytest-cov = "^2.4"
black = "=18.9b0"
